Bug 1135040: Optimize SIMD shifts in Ion; r=nbp
UltraBlame original commit: 3aca5bc01ad8cf0401b5bac86f07bfeb45a7b034

+++ js/src/jit/x86-shared/CodeGenerator-x86-shared.cpp
@@ -3107,12 +3107,9 @@ CodeGeneratorX86Shared::visitSimdShift(LSimdShift* ins)
     
     
     
-    
-    
-    
     const LAllocation* val = ins->value();
     if (val->isConstant()) {
-        int32_t c = ToInt32(val);
+        uint32_t c = uint32_t(ToInt32(val));
         if (c > 31) {
             switch (ins->operation()) {
               case MSimdShift::lsh:
